32-bit stack switching bug fix
R=ken
OCL=29412
CL=29412

+++ src/runtime/proc.c
@@ -705,7 +705,7 @@ newstack(void)
 
 #pragma textflag 7
 void
-sys·morestack(uint64 u)
+sys·morestack(uintptr u)
 {
 	while(g == m->g0) {
 		// very bad news
